Data analysis change: Use all data from RKI institute
In the past, we have ignored the last reported data point, because those
can sometimes be updated retrospectively, and thus start 'too low'. Too low
may encourage hope that the numbers are lower than they actually are, and thus
affect people's desire to stick to social distancing etc.

Within Germany, we are now entering the 'dance' regime of the 'hammer
and dance' approach to a pandemic. In this case, it is important to notice
new outbreaks quickly, and it seems country productive not to show the most
recent data, even if it may underestimate the final numbers initially.