Get rid of this file
The plan was to have an index file that could be used on binder
and in pelican to navigate all tutorials. However, on pelican, we
need the html extension in the links and on binder the ipynb extension.

Remove this for simplicity, and use the tag-tutorial.html page from pelican
as the index.

Each tutorial must have its own 'execute in binder' button, and the advanced
user will be able to jump from tutorial to tutorial once within binder
by opening the right notebook. Hopefully.